+++ idea/testData/refactoring/rename/renameJavaSyntheticIsPropertyByGetterRef/after/test.kt
@@ -1,0 +1,4 @@
+fun test(bean: Bean) {
+    bean.isProp2 = true
+    println(bean.isProp2)
+}