+++ idea/testData/refactoring/rename/renameJavaSyntheticPropertyByGetterSetterRef/before/test.kt
@@ -1,0 +1,5 @@
+fun test(bean: Bean) {
+    bean.prop = "a"
+    println(bean.prop)
+    bean./*rename*/prop += "a"
+}